rename fire module
